--- server/jue_jin_sign_in/info.json
@@ -1,6 +1,0 @@
-{
-  "aid": "XXXX",
-  "uuid": "XXXX",
-  "cookie": "XXXX",
-  "signature": "XXXX"
-}
